+++ package.json
@@ -2,7 +2,7 @@
   "name": "dt-exchange",
   "private": true,
   "scripts": {
-    "start": "npm run build -- --watch",
+    "start": "npm run build -- --watch --sourcemap",
     "build": "esbuild --bundle --jsx=automatic --loader:.png=dataurl --loader:.md=text ./src/bundle.tsx ./src/background --outdir=extension",
     "test": "jest",
     "typecheck": "tsc"
