+++ src/components/Item/index.ts
@@ -1,0 +1,2 @@
+export { Item } from "./Item"
+export { DEEMPHASIZE_OPTIONS, DeemphasizeOption } from "./Item"
