--- src/components/Store.css
@@ -1,79 +1,0 @@
-.item-container {
-  position: relative;
-}
-
-.item-title {
-  margin-bottom: 0.5em;
-}
-
-.item-already-owned {
-  opacity: 0.5;
-}
-
-.row:not(:first-child) {
-  border-top: 1px solid grey;
-  margin-top: 1em;
-  padding-top: 1em;
-}
-
-.info-items {
-  display: flex;
-  flex-direction: column;
-  /* margin-top: 1.2em; */
-
-  border-right-color: white;
-  border-right-width: 1px;
-  border-right-style: solid;
-  margin-right: 0.75em;
-}
-
-.info-item {
-  display: flex;
-  align-items: center;
-  margin: 0;
-  margin-right: 0.75em;
-  text-transform: capitalize;
-  /* flex: 1; */
-}
-.info-item img {
-  height: 3em;
-  float: left;
-  padding: 0.2em;
-}
-.stats {
-  display: grid;
-  grid-template-columns: repeat(3, minmax(0, 1fr));
-  grid-column-gap: 1em;
-  grid-row-gap: 1em;
-}
-.stat {
-}
-.stat-name {
-  color: rgb(189, 189, 189);
-}
-.stat-bar-row {
-  display: flex;
-  align-items: center;
-}
-.stat-p {
-  margin-right: 0.5em;
-}
-.stat-bar-outer {
-  width: 100%;
-  height: 0.7em;
-  border: 1px solid gold;
-  padding: 1px;
-}
-.stat-bar-inner {
-  position: relative;
-  background: gold;
-  height: 100%;
-}
-
-.perk {
-  display: flex;
-}
-
-.perk-rarity {
-  margin-right: 1em;
-}
